Workaround: Bug fixed: Victim identification. There are some problems in the victim identification state machine. It was re-programmed really simple as workaround here... Known bugs on the ramp (sometimes clears maze, finds no route possibility, detects victim on ramp...)

+++ Project/sys/main.h
@@ -92,7 +92,7 @@ extern int8_t timer_get_tast;
 extern int16_t timer_rdy_restart;
 extern int8_t timer_map_wall_r;
 extern int16_t timer_lop;
-extern int8_t timer_vic_ramp;
+extern int16_t timer_vic_ramp;
 
 extern int32_t incremental;
